% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_engine_coxph.R
\name{recur_fit.recur_engine_coxph}
\alias{recur_fit.recur_engine_coxph}
\alias{recur_predictfun.recur_engine_coxph}
\title{Fit a Cox PH with frailty engine}
\usage{
\method{recur_fit}{recur_engine_coxph}(engine, formula, data, ...)

\method{recur_predictfun}{recur_engine_coxph}(
  engine,
  newdata,
  eventtimes = NULL,
  type = c("survival", "cumhaz", "hazard"),
  log = FALSE
)
}
\arguments{
\item{engine}{A classed engine object created by [recur_engine()] and
usually fitted by [recur_fit()].}

\item{formula}{A one-sided model formula with predictors (` ~ covariates`).}

\item{data}{A `data.frame` containing variables referenced by `formula` and the following variables:
\describe{
  \item{time1:}{the beginning of time segments;}
  \item{time2:}{the end of time segments;}
  \item{id:}{subject identifiers;}
  \item{event:}{event indicator;}
  \item{terminal:}{indicator of terminal event}
  }
See \code{\link[reda]{Recur-class}} for more information}

\item{...}{Optional arguments forwarded to [survival::coxph()], such as `ties`, `robust`.}

\item{newdata}{A `data.frame` with covariates for all subjects to be predicted}

\item{eventtimes}{Numeric vector with the given event time `t` for each observation in `newdata`}

\item{type}{Character; one of `"survival"`, `"cumhaz"`, or `"hazard"`.}

\item{log}{Logical; whether the log-transformed value is required}
}
\value{
The updated engine object of class `"recur_engine_coxph"` with
  a fitted model embedded and any engine-specific metadata recorded.

A function(index, gaptimes).
}
\description{
S3 method for fitting a Cox proportional hazards model with frailty as the prediction
engine. Uses \pkg{survival} to fit the model and stores the
fitted object in `engine$model`.
}
\examples{
\dontrun{
eng <- recur_engine("coxph")
eng <- recur_fit(eng, Recur(time = time, id = patient.id, event = indicator) ~ Z1 + Z2,
                 data = SimulatedData, ties = "efron")
}
\dontrun{
survfun <- recur_predict(eng, newdata = SimulatedData[1:5, ], eventtimes = 2,
           type = "survival")
survfun(2, gaptimes =seq(0, 2, by = 0.1), type = "survival")
}
}
\seealso{
[survival::coxph()], [recur_predict.recur_engine_coxph()]
}
