% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_engine_known.R
\name{recur_predictfun.recur_engine_known}
\alias{recur_predictfun.recur_engine_known}
\title{Create a gap-time prediction function for a known model engine}
\usage{
\method{recur_predictfun}{recur_engine_known}(
  engine,
  newdata,
  eventtimes = NULL,
  type = c("survival", "cumhaz", "hazard"),
  log = FALSE
)
}
\arguments{
\item{engine}{A classed engine object created by [recur_engine()] and
usually fitted by [recur_fit()].}

\item{newdata}{A `data.frame` with covariates for all subjects to be predicted}

\item{type}{Character; one of `"survival"`, `"cumhaz"`, or `"hazard"`.}

\item{log}{Logical; whether the log-transformed value is required}
}
\value{
A function(index, gaptimes).
}
\description{
Create a gap-time prediction function for a known model engine
}
\details{
The following additional items need to be present in the engine:
\describe{
\item{`H0fun`}{a function giving the baseline gap-time cumulative hazard function}
\item{`coefs`}{optional numeric vector of coefficients used for the Cox model}
}
}
\examples{
\dontrun{
ke <- recur_engine("known",  H0fun = function(x) 0.5*x)
survfun <- recur_predictfun(ke, newdata = SimulatedData[1:5, ], eventtime = 2,
           type = "survival")
survfun(2, gaptimes =seq(0, 2, by = 0.1))
}
}
