% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_engine.R, R/recur_engine_coxph.R
\name{recur_fit}
\alias{recur_fit}
\alias{recur_fit.default}
\alias{recur_fit.recur_engine_coxph}
\title{Fit a model for a recurrent-event prediction engine}
\usage{
recur_fit(engine, formula, data, ...)

\method{recur_fit}{default}(engine, formula, data, ...)

\method{recur_fit}{recur_engine_coxph}(engine, formula, data, ...)
}
\arguments{
\item{engine}{A classed engine object returned by \code{recur_engine()}; e.g.,
class \code{recur_engine_coxph}.}

\item{formula}{A one-sided model formula with predictors (\code{~ covariates}).}

\item{data}{A data.frame containing variables referenced by \code{formula} and the following variables:
\describe{
  \item{time1}{the beginning of time segments;}
  \item{time2}{the end of time segments;}
  \item{id}{subject identifiers;}
  \item{event}{event indicator;}
  \item{terminal}{indicator of terminal event}
  }}

\item{...}{Additional arguments passed to the model.
Typically, it will be created by a call to \code{\link[reda]{Recur}}. See
\code{\link[reda]{Recur-class}} for more information.}
}
\value{
The same \code{engine} object, typically with its \code{model}
 field filled by a fitted object.
}
\description{
\code{recur_fit} is a generic to fit the underlying model required for predictions. The first argument must be a classed engine
created by \code{recur_engine()}

\code{recur_fit.recur_engine_coxph} is the S3 method for fitting a Cox proportional hazards model with frailty as the prediction
engine. It uses \pkg{survival} to fit the model and stores the
fitted object in \code{engine$model}.
}
\details{
The default method of \code{recur_fit} adds a \code{model} field
containing only the \code{formula} argument passed to it.
}
\examples{
eng <- recur_engine(name = "myname")
eng <- recur_fit(eng,  ~ X) # uses default method
eng

eng <- recur_engine("coxph")
eng <- recur_fit(eng, ~ Z.1 + Z.2,data = SimulatedData_recur, ties = "efron")
}
\seealso{
\code{\link{recur_engine}} for defining engines, and
\code{\link{recur_predictfun}} for an additional method that needs to be defined
}
