% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur.R_engine_coxph.R
\name{recur_fit.recur_engine_coxph}
\alias{recur_fit.recur_engine_coxph}
\title{Fit a Cox PH with frailty engine}
\usage{
\method{recur_fit}{recur_engine_coxph}(engine, formula, data, ...)
}
\arguments{
\item{engine}{A classed engine object returned by [recur_engine()]; e.g.,
class `"recur_engine_coxph"`.}

\item{formula}{A one-sided model formula with predictors (` ~ covariates`).}

\item{data}{A `data.frame` containing variables referenced by `formula` and the following variables:
\describe{
  \item{time1:}{the beginning of time segments;}
  \item{time2:}{the end of time segments;}
  \item{id:}{subject identifiers;}
  \item{event:}{event indicator;}
  \item{terminal:}{indicator of terminal event}
  }
See \code{\link[reda]{Recur-class}} for more information}

\item{...}{Additional arguments passed to the model}

\item{ties, }{robust, model, x, y Optional arguments forwarded to
[survival::coxph()] (if used internally).}
}
\value{
The updated engine object of class `"recur_engine_coxph"` with
  a fitted model embedded and any engine-specific metadata recorded.
}
\description{
S3 method for fitting a Cox proportional hazards model with frailty as the prediction
engine. Uses \pkg{survival} to fit the model and stores the
fitted object in `engine$model`.
}
\examples{
\dontrun{
eng <- recur_engine("coxph")
eng <- recur_fit(eng, Recur(time = time, id = patient.id, event = indicator) ~ Z1 + Z2,
                 data = SimulatedData, ties = "efron")
}
}
\seealso{
[survival::coxph()], [recur_predict.recur_engine_coxph()]
}
