% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_engine.R, R/recur_engine_coxph.R,
%   R/recur_engine_known.R
\name{recur_predictfun}
\alias{recur_predictfun}
\alias{recur_predictfun.recur_engine_coxph}
\alias{recur_predictfun.recur_engine_known}
\title{Create function that makes gap-time predictions for recurrent-event models}
\usage{
recur_predictfun(
  engine,
  newdata,
  eventtimes = NULL,
  type = c("survival", "cumhaz", "hazard"),
  log = FALSE
)

\method{recur_predictfun}{recur_engine_coxph}(
  engine,
  newdata,
  eventtimes = NULL,
  type = c("survival", "cumhaz", "hazard"),
  log = FALSE
)

\method{recur_predictfun}{recur_engine_known}(
  engine,
  newdata,
  eventtimes = NULL,
  type = c("survival", "cumhaz", "hazard"),
  log = FALSE
)
}
\arguments{
\item{engine}{A classed engine object created by \code{\link{recur_engine}} and
usually fitted by \code{\link{recur_fit}}.}

\item{newdata}{A data.frame with covariates for all subjects to be predicted}

\item{eventtimes}{Numeric vector with the given event time \eqn{t} for
each observation in \code{newdata}.}

\item{type}{Character; one of "survival", "cumhaz", or "hazard".}

\item{log}{Logical; whether the log-transformed value is required}
}
\value{
a function that will make predictions on \code{newdata} from the following inputs
 \describe{
  \item{index}{Integer; indicates the observation number from \code{newdata} for prediction.}
  \item{gaptimes}{Numeric vector of gap-times \eqn{y}.}
 }
}
\description{
\code{recur_predictfun} is a generic to produce time-indexed predictions required by
the estimators in \code{\link{estimate_end}}: the survival  \eqn{S(y|X, t)},
 cumulative hazard \eqn{H(y|X, t)}, or hazard \eqn{h(y|X, t)} for gap-time \eqn{y} after
 an event at time \eqn{t} for an observation with covariates \eqn{X}.

\code{recur_predictfun.recur_engine_coxph} is the S3 method for
creating a prediction function based on a fitted Cox proportional hazards model
with frailty.

\code{recur_predictfun.recur_engine_known} is the S3 method for
creating a prediction function for a Cox regression model with a known baseline function.
}
\details{
For \code{recur_predictfun.recur_engine_known} the following additional items need to
be present in the engine:
\describe{
\item{H0fun}{a function giving the baseline gap-time cumulative hazard function}
\item{coefs}{optional numeric vector of coefficients used for the Cox model}
}
}
\examples{
eng <- recur_engine(name = "coxph")
eng <- recur_fit(eng,  ~ Z.1 + Z.2, data = SimulatedData_recur)
predfun <- recur_predictfun(eng, newdata = SimulatedData_recur[1:5, ], eventtimes = 2,
                         type = "survival", log=TRUE)
predfun(2, gaptimes =seq(0, 2, by = 0.1))

eng <- recur_engine("coxph") |>
  recur_fit(formula= ~ Z.1 + Z.2,data = SimulatedData_recur)
nd <- data.frame(id = 1:4, Z.1 = c(0,0,1,1), Z.2 = c(0, 0.1, 0.2, 0.4))
survfun <- recur_predictfun(eng, newdata = nd, eventtimes = 2,
           type = "survival")
# predict for row 2 in 'nd'
survfun(2, gaptimes =seq(0, 2, by = 0.1))

ke <- recur_engine("known",  H0fun = function(x) 0.5*x) |>
   recur_fit(~1)  # no predictors
survfun <- recur_predictfun(ke, newdata = SimulatedData[1:5, ], eventtimes = 2,
           type = "survival")
survfun(2, gaptimes =seq(0, 2, by = 0.1))

}
\seealso{
\code{\link{recur_engine}} for defining engines, and
\code{\link{recur_fit}} for an additional method that needs to be defined
}
