% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur.R_engine_coxph.R
\name{recur_predictfun.recur_engine_coxph}
\alias{recur_predictfun.recur_engine_coxph}
\title{Create a gap-time prediction funcition for a Cox PH with frailty engine}
\usage{
\method{recur_predictfun}{recur_engine_coxph}(
  engine,
  newdata,
  eventtimes = NULL,
  type = c("survival", "cumhaz", "hazard"),
  log = FALSE
)
}
\arguments{
\item{engine}{A classed engine object created by [recur_engine()] and
usually fitted by [recur_fit()].}

\item{newdata}{A `data.frame` with covariates for all subjects to be predicted}

\item{type}{Character; one of `"survival"`, `"cumhaz"`, or `"hazard"`.}

\item{log}{Logical; whether the log-transformed value is required}
}
\value{
A function(index, gaptimes).
}
\description{
Create a gap-time prediction funcition for a Cox PH with frailty engine
}
\examples{
\dontrun{
survfun <- recur_predict(eng, newdata = SimulatedData[1:5, ], eventtime = 2,
           type = "survival")
survfun(2, gaptimes =seq(0, 2, by = 0.1), type = "survival")
}
}
