% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_recur_end.R
\name{sim_recur_end}
\alias{sim_recur_end}
\title{Simulate recurrent event data with unobserved terminating event}
\usage{
sim_recur_end(
  n,
  lambda_d,
  lambda_r,
  sigma2,
  lambda_c,
  C_min = 0,
  cov_fun = NULL,
  beta = NULL,
  censored.ok = TRUE,
  ...
)
}
\arguments{
\item{n}{integer, sample size of independent subjects}

\item{lambda_d}{numeric value > 0, rate of exponentially distributed
terminating event}

\item{lambda_r}{numeric value > 0, baseline rate of recurrent event process}

\item{sigma2}{numeric value >= 0, variance of gamma-distributed random effect
for the recurrent event process}

\item{lambda_c}{numeric value > 0, rate of censoring distribution after minimal
follow-up time \code{C_min}}

\item{C_min}{numeric value >=0, minimal follow-up time without censoring}

\item{cov_fun}{function of the form \code{function(n,...)} that generates a
model matrix of covariates for \code{n} subjects. Set to NULL if no covariates
are used.}

\item{beta}{numeric vector with the coefficients of the covariates generated
by \code{cov_fun}. Ignored if \code{cov_fun} is NULL.}

\item{censored.ok}{logical, indicating whether subjects with first event censored
(ie `nevents=0`) should be retained}

\item{...}{optional parameters passed to \code{cov_fun}}
}
\value{
A data frame with the simulated event times and the values
of the unobservable underlying variables:
\describe{
  \item{patient.id}{int Patient ID}
  \item{nevents}{int Number of observed recurrent events for the patient}
  \item{time}{dbl Observation time of a recurrent event or censoring}
  \item{indicator}{int Event indicator. 1 - recurrent event, 0 - censoring}
  \item{recurrence_end}{dbl _Unobservable_ end-time of the recurrent event process}
  \item{C}{dbl Censoring time}
  \item{xi}{dbl _Unobservable_ subject-specific frailty of the
            recurrent event process}
  \item{Z.1, Z.2, ...}{the values of the covariates generated by \code{cov_fun}}
  }
}
\description{
Simulate recurrent event data with unobserved terminating event
}
\examples{
simd <- sim_recur_end(n=10, lambda = 1, lambda_r = 5, sigma2=0.1,
   cov_fun = function(n)cbind(rbinom(n, size=1, prob = 0.5),
                              rnorm(n)),
   beta = c(0,0.2), lambda_c = 1, C_min = 1)

}
