% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfun.R
\name{plot.survfun}
\alias{plot.survfun}
\alias{lines.survfun}
\alias{predict.survfun}
\alias{quantile.survfun}
\alias{median.survfun}
\title{Methods for 'survfun' objects}
\usage{
\method{plot}{survfun}(
  x,
  conf.int = FALSE,
  ylim = c(0, 1),
  conf.lty = 2,
  conf.col = 1,
  do.points = FALSE,
  ...
)

\method{lines}{survfun}(x, conf.int = FALSE, conf.lty = 2, conf.col = 1, ...)

\method{predict}{survfun}(object, times = NULL, conf.int = FALSE, ...)

\method{quantile}{survfun}(x, probs = c(0.25, 0.5, 0.75), conf.int = TRUE, ...)

\method{median}{survfun}(x, na.rm = FALSE, conf.int = TRUE, ...)
}
\arguments{
\item{x}{object of class 'survfun'}

\item{conf.int}{logical, whether confidence limits / intervals should be shown.
Defaults to \code{FALSE}.}

\item{ylim}{numeric vector of length 2 defining the y limits of the plot. Defaults to \code{c(0,1)}.}

\item{conf.lty}{integer, linetype for plotted confindence limits}

\item{conf.col}{color of plotted confindence limits}

\item{do.points}{logical, specifying whether points should be plotted at the closed end
of each step. Defaults to \code{FALSE}.}

\item{...}{additional parameters, only used for the \code{plot} method, where they
are passed to \code{\link[stats]{plot.stepfun}}}

\item{object}{object of class 'survfun'}

\item{times}{numeric vector of times (x-values) for which prediction is requested.
Defaults to all jump-points of the function.}

\item{probs}{numeric vector of probabilities. Defaults to all quartiles.}

\item{na.rm}{logical, required as part of the generic for \code{median}. Ignored.}
}
\description{
List of methods that apply to \code{survfun} objects created by the
\code{\link{estimate_end}} function.
}
