% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_end.R
\name{estimate_end}
\alias{estimate_end}
\title{Estimate the distribution of the ending event of a recurrent event process}
\usage{
estimate_end(
  formula,
  method = c("naive", "threshold", "quantile", "NPMLE"),
  threshold = 0,
  quantile = 0.95,
  data,
  subset,
  na.action,
  verbose = FALSE,
  known_recur = NULL,
  IPSW = FALSE
)
}
\arguments{
\item{formula}{a two-sided formula with a \code{\link[reda]{Recur}} object specifying the
recurrent event process on the left-hand side, and the predictors of its intensity
on the right-hand side. Origin and terminal events in \code{Recur} are not respected.}

\item{method}{character string specifying the estimation method}

\item{threshold}{optional numeric value, specifies the threshold for the "threshold" method.
Defaults to 0, which is equivalent to the "naive" method.}

\item{quantile}{optional numeric value between 0 and 1 (exclusive) specifying the
quantile used for the "quantile" method.}

\item{data}{an optional data frame containing the variables in the formula}

\item{subset}{an optional vector specifying a subset of observations to be used}

\item{na.action}{a function which indicates what should happen when the data
contains NAs. The default is set by the na.action setting of \code{options},
which in turn defaults to \code{na.omit}.}

\item{verbose}{logical value, if TRUE then information is displayed during computation}

\item{known_recur}{optional list specifying the known recurrence survival distribution
function for each subject. It should have two components: \code{S0} - the baseline survival
function, and \code{coefs} - the coefficients of the recurrence model specified in \code{formula}}

\item{IPSW}{logical, indicating whether inverse-probability of selection weights based on
no censoring before first event time should be used}
}
\value{
a list with components \code{fit} - an object of class \link[stats]{stepfun} with the
estimated survival distribution, \code{method} - the method used, and additional
information based on the method used
}
\description{
Estimate the distribution of the ending event of a recurrent event process
}
