% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_end.R
\name{estimate_end}
\alias{estimate_end}
\title{Estimate the distribution of the ending event of a recurrent event process}
\usage{
estimate_end(
  formula,
  method = c("naive", "threshold", "quantile", "NPMLE"),
  threshold = 0,
  quantile = 0.95,
  data,
  subset,
  na.action,
  verbose = FALSE,
  engine = recur_engine("coxph"),
  known_recur = NULL,
  IPSW = FALSE,
  bootCI = FALSE,
  conf.level = 0.95,
  bootB = 100
)
}
\arguments{
\item{formula}{a two-sided formula with a \code{\link[reda]{Recur}} object specifying the
recurrent event process on the left-hand side, and the predictors of its intensity
on the right-hand side. Origin values in \code{Recur} are not respected, but terminal
events are accounted for.}

\item{method}{character string specifying the estimation method}

\item{threshold}{optional numeric value, specifies the threshold for the "threshold" method.
Defaults to 0, which is equivalent to the "naive" method.}

\item{quantile}{optional numeric value between 0 and 1 (exclusive) specifying the
quantile used for the "quantile" method.}

\item{data}{an optional data frame containing the variables in the formula}

\item{subset}{an optional vector specifying a subset of observations to be used}

\item{na.action}{a function which indicates what should happen when the data
contains NAs. The default is set by the na.action setting of \code{options},
which in turn defaults to \code{na.omit}.}

\item{verbose}{logical value, if TRUE then information is displayed during computation}

\item{engine}{an object of class `recur_engine` for fitting the recurrent event model}

\item{known_recur}{_deprecated_ optional list specifying the known recurrence survival distribution
function for each subject. It should have two components: \code{S0} - the baseline survival
function, and \code{coefs} - the coefficients of the recurrence model specified in \code{formula}}

\item{IPSW}{logical, indicating whether inverse-probability of selection weights based on
no censoring before first event time should be used}

\item{bootCI}{logical, indicating whether pointwise bootstrap confidence intervals should
be computed. Use carefully for the "NPMLE" method, as it may take long.}

\item{conf.level}{numeric, confidence level for pointwise confidence interval}

\item{bootB}{integer, number of bootstrap replicates}
}
\value{
an object of class \code{\link{survfun}} with the estimated survival distribution
}
\description{
Estimate the distribution of the ending event of a recurrent event process
}
\examples{
# Naive estimator
res_n <- estimate_end(Recur(time=time, id=patient.id, event=indicator) ~ 1,
         method="naive", data=SimulatedData)
median(res_n)
# Threshold estimator using the 99th quantile of the recurrent event distribution
res_q <- estimate_end(Recur(time=time, id=patient.id, event=indicator) ~ 1,
                     method="quantile", data=SimulatedData, quantile=0.99)
quantile(res_q)
# NPMLE estimator, adjusting for covariates in the recurrent event estimation
res_np <- estimate_end(Recur(time=time, id=patient.id, event=indicator) ~ Z.1 + Z.2,
                      method="NPMLE", data=SimulatedData)
plot(res_np)
# request bootstrap confidence intervals (B should be larger in real application)
res_npb <- estimate_end(Recur(time=time, id=patient.id, event=indicator) ~ Z.1 + Z.2,
                      method="NPMLE", bootCI=TRUE, bootB=5, data=SimulatedData)
plot(res_npb, conf.int=TRUE)
}
\seealso{
[median()], [quantile()], [predict()], and [plot()] for
working with the objects of class 'survfun'
}
