% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_engine.R
\name{recur_fit}
\alias{recur_fit}
\title{Fit a model for a recurrent-event prediction engine}
\usage{
recur_fit(engine, formula, data, ...)
}
\arguments{
\item{engine}{A classed engine object returned by [recur_engine()]; e.g.,
class `"recur_engine_coxph"`.}

\item{formula}{A one-sided model formula with predictors (` ~ covariates`).}

\item{data}{A `data.frame` containing variables referenced by `formula` and the following variables:
\describe{
  \item{time1:}{the beginning of time segments;}
  \item{time2:}{the end of time segments;}
  \item{id:}{subject identifiers;}
  \item{event:}{event indicator;}
  \item{terminal:}{indicator of terminal event}
  }
See \code{\link[reda]{Recur-class}} for more information}

\item{...}{Additional arguments passed to the model}
}
\value{
The same `engine` object, typically with its `model` field filled
  by a fitted object.
}
\description{
Generic to fit the underlying model required for predictions. The first argument must be a classed engine
created by [recur_engine()].
}
\details{
The default method adds a `model` field containing only the `formula`.
}
\examples{
\dontrun{
eng <- recur_engine("coxph")
eng <- recur_fit(eng,  ~ Z1 + Z2, data = SimulatedData)
}
}
\seealso{
[recur_predict()], [estimate_end()]
}
