% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_engine.R
\name{recur_predictfun}
\alias{recur_predictfun}
\title{Create function that makes gap-time predictions for recurrent-event models}
\usage{
recur_predictfun(
  engine,
  newdata,
  eventtimes = NULL,
  type = c("survival", "cumhaz", "hazard"),
  log = FALSE
)
}
\arguments{
\item{engine}{A classed engine object created by [recur_engine()] and
usually fitted by [recur_fit()].}

\item{newdata}{A `data.frame` with covariates for all subjects to be predicted}

\item{type}{Character; one of `"survival"`, `"cumhaz"`, or `"hazard"`.}

\item{log}{Logical; whether the log-transformed value is required}

\item{eventtime}{Numeric vector with the given event time `t` for each observation in `newdata`}
}
\value{
a function that will make predictions on `newdata` from the following inputs
 \describe{
  \item{index}{Integer; indicates the observation number from `newdata` for prediction.}
  \item{gaptimes}{Numeric vector of gap-times `y`.}
 }
}
\description{
Generic to produce time-indexed predictions required by the estimators: survival `S(y|X, t)`,
 cumulative hazard `H(y|X, t)`, or hazard `h(y|X, t)`, for gap-time `y` after
 an event at time `t` for an observation with covariates `X`.
}
\examples{
\dontrun{
predfun <- recur_predict(eng, newdata = SimulatedData[1:5, ], eventtime = 2,
                         type = "survival", log=TRUE)
predfun(2, gaptimes =seq(0, 2, by = 0.1))
}
}
\seealso{
[recur_fit()], [estimate_end()]
}
