% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_recur_end.R
\name{sim_recur_end}
\alias{sim_recur_end}
\title{Simulate recurrent event data with unobserved terminating event}
\usage{
sim_recur_end(
  n,
  lambda_d,
  lambda_r,
  sigma2,
  cov_fun,
  beta,
  lambda_c,
  C_min = 0,
  ...
)
}
\arguments{
\item{n}{integer, sample size of independent subjects}

\item{lambda_d}{numeric value > 0, rate of exponentially distributed
terminating event}

\item{lambda_r}{numeric value > 0, baseline rate of recurrent event process}

\item{sigma2}{numeric value >= 0, variance of gamma-distributed random effect
for the recurrent event process}

\item{cov_fun}{function of the form \code{function(n,...)} that generates a
model matrix of covariates for \code{n} subjects}

\item{beta}{numeric vector with the coefficients of the covariates generated
by \code{cov_fun}}

\item{lambda_c}{numeric value > 0, rate of censoring distribution after minimal
follow-up time \code{C_min}}

\item{C_min}{numeric value >=0, minimal follow-up time without censoring}

\item{...}{optional parameters passed to \code{cov_fun}}
}
\description{
Simulate recurrent event data with unobserved terminating event
}
