% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfun.R
\name{survfun}
\alias{survfun}
\alias{plot.survfun}
\alias{lines.survfun}
\alias{predict.survfun}
\alias{quantile.survfun}
\alias{median.survfun}
\title{Methods for 'survfun' objects}
\usage{
\method{plot}{survfun}(
  x,
  conf.int = FALSE,
  ylim = c(0, 1),
  conf.lty = 2,
  conf.col = 1,
  do.points = FALSE,
  main = "",
  xlab = "Time",
  ylab = "Survival function",
  ...
)

\method{lines}{survfun}(x, conf.int = FALSE, conf.lty = 2, conf.col = 1, ...)

\method{predict}{survfun}(object, times = NULL, conf.int = FALSE, ...)

\method{quantile}{survfun}(x, probs = c(0.25, 0.5, 0.75), conf.int = TRUE, ...)

\method{median}{survfun}(x, na.rm = FALSE, conf.int = TRUE, ...)
}
\arguments{
\item{x}{object of class 'survfun'}

\item{conf.int}{logical, whether confidence limits / intervals should be shown.
Defaults to \code{FALSE}.}

\item{ylim}{numeric vector of length 2 defining the y limits of the plot. Defaults to \code{c(0,1)}.}

\item{conf.lty}{integer, linetype for plotted confindence limits}

\item{conf.col}{color of plotted confindence limits}

\item{do.points}{logical, specifying whether points should be plotted at the closed end
of each step. Defaults to \code{FALSE}.}

\item{main}{character main title of the plot.}

\item{xlab, ylab}{character labels for the x and y axis.}

\item{...}{additional parameters, only used for the \code{plot} method, where they
are passed to \code{\link[stats]{plot.stepfun}}}

\item{object}{object of class 'survfun'}

\item{times}{numeric vector of times (x-values) for which prediction is requested.
Defaults to all jump-points of the function.}

\item{probs}{numeric vector of probabilities. Defaults to all quartiles.}

\item{na.rm}{logical, required as part of the generic for \code{median}. Ignored.}
}
\description{
List of methods that apply to \code{survfun} objects created by the
\code{\link{estimate_end}} function.
}
\examples{
res_q <- estimate_end(Recur(time=time, id=patient.id, event=indicator) ~ 1,
                     method="quantile", data=SimulatedData, quantile=0.99)
res_n <- estimate_end(Recur(time=time, id=patient.id, event=indicator) ~ 1,
                     method="naive", data=SimulatedData)
# plot estimated survival function
plot(res_q, conf.int=TRUE, col="blue", conf.col="navy", conf.lty=3)
lines(res_n, col="darkred")

# get quantiles
median(res_q)
quantile(res_q, probs=c(0.1, 0.9))

# values at specific time-point
predict(res_q, times = c(0.5, 1, 1.5))

}
